LDR change to uint16_t

+++ bruh_mqtt_multisensor_github/bruh_mqtt_multisensor_github.ino
@@ -486,7 +486,8 @@ void loop() {
     }
 
 
-    int newLDR = lightMeter.readLightLevel();
+    //int newLDR = lightMeter.readLightLevel();
+    uint16_t newLDR = lightMeter.readLightLevel();
 
     if (checkBoundSensor(newLDR, LDR, diffLDR)) {
       LDR = newLDR;
